# storage/docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "CRC",
        "description": "This is a sample server Petstore server.  You can find out more about Swagger at <a href=\"http://swagger.io\">http://swagger.io</a> or on irc.freenode.net, #swagger.  For this sample, you can use the api key \"special-key\" to test the authorization filters",
        "termsOfService": "http://helloreverb.com/terms/",
        "contact": {
            "email": "zaro.megagenta@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8000",
    "basePath": "/api",
    "schemes": [
        "http"
    ],
    "paths": {
        "/classTypes": {
            "get": {
                "tags": [
                    "ClassType"
                ],
                "summary": "Find all instances of the model matched by filter from the data source.",
                "operationId": "ClassTypesController.find",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "data",
                        "in": "body",
                        "description": "Model instance data",
                        "required": true,
                        "schema": {
                            "required": [
                                "id"
                            ],
                            "$ref": "#/definitions/ClassType"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/PaginatedClassType"
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "ClassType"
                ],
                "summary": "Create a new instance of the model and persist it into the data source.",
                "operationId": "ClassTypesController.create",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "data",
                        "in": "body",
                        "description": "Model instance data",
                        "required": true,
                        "schema": {
                            "required": [
                                "id"
                            ],
                            "$ref": "#/definitions/ClassType"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/ClassType"
                        }
                    },
                    "422": {
                        "description": "Invalid input",
                        "schema": {
                            "$ref": "#/definitions/errors/CreateClassTypeRequest"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "PaginatedClassType": {
            "properties": {
                "total": {
                    "type": "integer",
                    "format": "int64"
                },
                "per_page": {
                    "type": "integer",
                    "format": "int64"
                },
                "current_page": {
                    "type": "integer",
                    "format": "int64"
                },
                "last_page": {
                    "type": "integer",
                    "format": "int64"
                },
                "name": {
                    "type": "string",
                    "minLength": 3
                }
            }
        },
        "ClassType": {
            "required": [
                "name"
            ],
            "properties": {
                "id": {
                    "type": "integer",
                    "format": "int64"
                },
                "name": {
                    "type": "string",
                    "minLength": 3
                }
            }
        },
        "errors/CreateClassTypeRequest": {
            "properties": {
                "name": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        }
    },
    "externalDocs": {
        "description": "Find out more about Swagger",
        "url": "http://swagger.io"
    }
}